+++ CppAsm/Memory.cpp
@@ -12,19 +12,19 @@ Memory::~Memory()
 	delete [] memory;
 }
 
-char *Memory::operator[](int index)
+char *Memory::operator[](size_t index)
 {
 	return &memory[index];
 }
 
-void Memory::read(int from, byte * buffer, int count)
+void Memory::read(size_t from, byte * buffer, size_t count)
 {
 	for(int i=0;i<count;i++)
 		buffer[i] = memory[from + i];
 	
 }
 
-int Memory::getSize()
+size_t Memory::getSize()
 {
 	return size;
 }
